Update app.py
saves quantized model to disk when using "python app.py --int8", then on next load it will load the quantized model from the disk so that it no longer has to quantize it every launch. You still need to use "--int8" when running to tell it to use the quantized version.